@@ -1,14 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
+import ReactDOM from "react-dom/client";
 import App from "./App";
+import "./index.css";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
     <React.StrictMode>
-        <div className="app-wrapper">
-            <App />
-        </div>
-    </React.StrictMode>,
-    document.getElementById("root")
+        <App />
+    </React.StrictMode>
 );
-
