@@ -2,24 +2,9 @@ const mongoose = require("mongoose");
 
 module.exports = async () => {
     try {
-        const connectionParams = {
-            // user: process.env.MONGO_USERNAME,
-            // pass: process.env.MONGO_PASSWORD,
-            useNewUrlParser: true,
-            // useCreateIndex: true,
-            useUnifiedTopology: true,
-        };
-        const useDBAuth = process.env.USE_DB_AUTH || false;
-        if(useDBAuth){
-            connectionParams.user = process.env.MONGO_USERNAME;
-            connectionParams.pass = process.env.MONGO_PASSWORD;
-        }
-        await mongoose.connect(
-           process.env.MONGO_CONN_STR,
-           connectionParams
-        );
+        await mongoose.connect(process.env.MONGO_CONN_STR);
         console.log("Connected to database.");
     } catch (error) {
-        console.log("Could not connect to database.", error);
+        console.error("Could not connect to database:", error);
     }
 };
