@@ -2,7 +2,7 @@ const Task = require("../models/task");
 const express = require("express");
 const router = express.Router();
 
-router.post("/", async (req, res) => {
+router.post("/", async (req, res) => {    
     try {
         const task = await new Task(req.body).save();
         res.send(task);
