@@ -1,6 +1,7 @@
 import axios from "axios";
-const apiUrl = process.env.REACT_APP_BACKEND_URL //"http://localhost:8080/api/tasks";
-console.log(apiUrl)
+
+const apiUrl = process.env.REACT_APP_BACKEND_URL // "http://localhost:3500/api/tasks";
+
 export function getTasks() {
     return axios.get(apiUrl);
 }
@@ -10,9 +11,9 @@ export function addTask(task) {
 }
 
 export function updateTask(id, task) {
-    return axios.put(apiUrl + "/" + id, task);
+    return axios.put(`${apiUrl}/${id}`, task);
 }
 
 export function deleteTask(id) {
-    return axios.delete(apiUrl + "/" + id);
+    return axios.delete(`${apiUrl}/${id}`);
 }
